update io_test and remove redundant gen_ranges
